upgrade angular-ui-bootstrap to fix an IE problem
Specifically, upgrade angular-ui-bootstrap to 0.3.0 and angular.js to 1.0.8.

This was done in production a while back, but apparently I didn't commit the changes to this repo.
I don't remember the problem exactly; I think there was a bad css class that was being applied that caused the modal to fail to disappear.
Upgrading fixed the bug.