switch from session cookies to explicitly-expired cookies
Chrome and Firefox both have changed how they handle session cookies, and browser restarts do not necessarily clear them anymore. As a result, the popup was being suppressed far longer than we originally intended.

This switches to persistent cookies that explicitly expire in 24 hours.

This changes the cookie name so that pre-existing session cookies will not be problematic.

This also adds a better error message when there is a server failure after the user clicks 'Agree'.

It would be better to switch from cookies to local storage or something, since this value doesn't need to be sent to the server. However, we'd have to add some manual expiration logic, and it's all a bit more work than I'm capable of doing quickly.