fix text/plain handling
This was causing a javascript error when someone clicked one of the buttons, but it didn't trigger an error message and so no one has been complaining. Their response was being saved, and the cookie was being saved on the next request, so the behavior impact was minimal.

Still, it's nice to avoid errors.